+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * All steps now have `required_pkgs()` methods.
 
+* `step_collapse_stringdist()` has gained `method` and `options` arguments to allow for different types of string distance calculations. (#152)
+
 # embed 1.0.0
 
 * `step_collapse_cart()` can pool a predictor's factor levels using a tree-based method.